pyproject.toml
[project]
name = "fastapi-tarea3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "cheve17", email = "cheve.iteso@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.117.1",
    "pydantic-settings>=2.11.0",
    "python-dotenv>=1.1.1",
    "uvicorn[standard]>=0.37.0",
]

[project.scripts]
fastapi-tarea3 = "fastapi_tarea3:main"

[build-system]
requires = ["uv_build>=0.8.14,<0.9.0"]
build-backend = "uv_build"
